refactor vibe coded garbage
